+++ lib/bloc_ease.dart
@@ -6,6 +6,5 @@ export 'src/bloc_ease_state.listener.dart';
 export 'src/state_widgets.provider.dart';
 export 'src/bloc_ease_combined_state.listener.dart';
 export 'src/bloc_ease_combined_state.builder.dart';
-export 'src/bloc_ease_cubit.dart';
 export 'src/bloc_ease_states.dart';
-export 'src/bloc_ease_cached_cubit.dart';
+export 'src/bloc_ease_state_cache.mixin.dart';